OVERFLOW-46 Updated Enemy class to include a description for created enemies-- updated contructor to include this
Game.cpp now gives a decription of the enemy to the player before a battle is begun.

GameManager method GameBattleManager now includes a check to see if the player is in the first battle with the dust golem-- if yes, it prints out the prompt for the tutorial battle

Player default damage updated to be 2, so that deflection can be showcased to work. it was 1 before which caused some issues with the way deflection works